Moved tests so TestUniject.cs only depends on core uniject classes
