+++ src/test/tests/src/Tests/TestExamplePhysics.cs
@@ -1,0 +1,18 @@
+using Ninject;
+using NUnit.Framework;
+using System;
+using UnityEngine;
+
+namespace Tests {
+
+    [TestFixture]
+    public class TestExamplePhysics : BaseInjectedTest {
+
+        [Test]
+        public void testRaycastScene() {
+            kernel.Get<ScanningLaser>();
+            step(10);
+        }
+    }
+}
+
