+++ src/test/tests/src/Tests/TestUniject.cs
@@ -136,17 +136,6 @@ namespace Tests {
         }
 
         /// <summary>
-        /// Our testable example should create exactly two game objects; one for the component itself,
-        /// and another for its injected prefab.
-        /// </summary>
-        [Test]
-        public void testTestableSceneObjectCreationCount() {
-            kernel.Get<TestableExample>();
-            step(1); // Must step a frame to ensure our test updatable manager tracks all objects.
-            Assert.AreEqual(2, kernel.Get<TestUpdatableManager>().Count);
-        }
-
-        /// <summary>
         /// Injected raw <c>TestableGameObject</c>.
         /// </summary>
         [Test]
@@ -273,12 +262,6 @@ namespace Tests {
             mockPhysics.Setup(x => x.Raycast(It.IsAny<Vector3>(), It.IsAny<Vector3>(), It.IsAny<float>(), It.IsAny<int>())).Returns(true);
 
             Assert.IsTrue(kernel.Get<IPhysics>().Raycast(Vector3.zero, Vector3.one, float.MaxValue, 0));
-        }
-
-        [Test]
-        public void testRaycastScene() {
-            kernel.Get<ScanningLaser>();
-            step(10);
         }
     }
 }
