+++ src/test/tests/test.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -53,6 +53,7 @@
     <Compile Include="src\Impl\mock\MockLayerMask.cs" />
     <Compile Include="src\Impl\mock\MockTime.cs" />
     <Compile Include="src\Tests\TestExampleGameObject.cs" />
+    <Compile Include="src\Tests\TestExamplePhysics.cs" />
     <Compile Include="src\Tests\TestUniject.cs" />
     <Compile Include="src\Impl\framework\FileVerifyingResourceProvider.cs" />
     <Compile Include="src\Tests\TestCollisions.cs" />
